index.php: count number of subscribers per group
Count the number of users into a group, and add the number on the right of
each group. Allows you to know the number of users per group at a glance
(i.e. you don't need to inspect each group anymore).

Trust that $row[0] is an integer, since we know it's an integer from
the database schema.

While there, make it more explicit that we sanitize the title that was
stored into the database.